+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 ### HEAD
 
+### 0.0.2 (September 12, 2019)
+
+  * Added the method $().firstChild(),
+  * Added a root parent to $(),
+  * Added the method trigger to $(),
+  * Added the method View.restore to remove all traces of a View Component from the DOM,
+  * ...,
+
+
 ### 0.0.1 (September 10, 2019)
 
   * Fixed and issue that prevents animate duration to be expressed in millisecond,